% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_query_wrappers.R
\name{gemini_caller}
\alias{gemini_caller}
\title{GEMINI query wrapper}
\usage{
gemini_caller(gemini_db, test = "autosomal_recessive",
  filter = "aaf < 0.1 AND aaf_esp_all < 0.01 AND \\n                        aaf_1kg_all < 0.01 AND af_exac_all < 0.01 AND \\n                        (is_coding=1 OR is_splicing=1 OR impact_severity='HIGH') \\n                        AND filter is NULL",
  min_gq = 20, families = NA, ...)
}
\arguments{
\item{gemini_db}{is the name of your GEMINI database (with path, if necessary)}

\item{test}{is the name of the GEMINI sub command to call. Accepted tests are 
autosomal_dominant, autosomal_recessive, comp_hets, mendel_errors, x_linked_de_novo,
x_linked_dominant, x_linked_recessive}

\item{filter}{if you want to change the default filtering criteria}

\item{min_gq}{minimum genotype quality (default is set at 20)}

\item{families}{family name that GEMINI will use to identify proband, mother, and father}

\item{...}{add other GEMINI commands not listed here}
}
\value{
None
}
\description{
Calls GEMINI's built in tools to retrieve proband/trio specific variants. Returns a tibble of all of the columns available in your GEMINI database.
}
\examples{
gemini_caller('/path/to/your/gemini.db', 'autosomal_dominant', families = 'fam007')
}
