% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_query_wrapper.R
\name{gemini_query_wrapper}
\alias{gemini_query_wrapper}
\title{GEMINI query wrapper}
\usage{
gemini_query_wrapper(gemini_db, test_name = "CUSTOM1", output = NA, ...)
}
\arguments{
\item{gemini_db}{is the name of your GEMINI database (with path, if necessary)}

\item{test_name}{The identifying name given to your query. Will be used in the 
interative document to filter query/GEMINI sub command types}

\item{output}{By default the data frame will be returned. If you give path/name.tsv
a tab separate file will be written}

\item{...}{Your GEMINI command, past `gemini query -q`}
}
\value{
None
}
\description{
A barebones GEMINI wrapper which passes on your GEMINI command to `gemini query 
--header -q` on the command line. I imagine this will mostly be used for users 
with already functional `gemini query --header -q` commands. You \strong{must} 
escape (append \) your quote marks.
}
\examples{
gemini_query_wrapper('/path/to/your/gemini.db', ... = "\\"SELECT * FROM variants WHERE (aaf_esp_all < 0.01 AND IMPACT_SO LIKE '\%STOP\%' AND filter is NULL) LIMIT 20\\"")
}
