% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_stat_wrapper.R
\name{gemini_stat_wrapper}
\alias{gemini_stat_wrapper}
\title{GEMINI stat wrapper}
\usage{
gemini_stat_wrapper(gemini_db, ID, is_family = "yes")
}
\arguments{
\item{gemini_db}{is the name of your GEMINI database (with path, if necessary)}

\item{ID}{is the family name or sample name to pull stats from}

\item{is_family}{Defaults to `yes`. The GEMINI databases's ped will be pulled 
to find the sample names associated with the family_id. If this is a sample 
name, change to 'no'.}
}
\value{
None
}
\description{
Uses GEMINI database to extracts some useful stats from the 
`gemini stats` command. Returns list of stats info that 
\code{\link{knit_see_gem}} uses to build the stats tab.
}
\examples{
gemini_stat_wrapper('/path/to/gemini.db', 'Sample007', is_family = 'no')
}
