% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_see_gem.R
\name{knit_see_gem}
\alias{knit_see_gem}
\title{Knit SeeGEM with peddy stats}
\usage{
knit_see_gem(rmd = system.file("rmd/document_template.Rmd", package =
  "SeeGEM"), output_file = "SeeGEM_document.html",
  output_directory = getwd(),
  GEMINI_data = system.file("extdata/GEMINI_data.Rdata", package =
  "SeeGEM"), sample_name = NA, title = "SeeGEM Test Report",
  peddy_path_prefix = paste0(system.file("extdata/", package = "SeeGEM"),
  "/SEE_GEM_PEDDY"), peddy_id = c("1045", "1046", "1265"),
  decorate = TRUE, skip_stats = "no")
}
\arguments{
\item{rmd}{Path to a custom R markdown file. A default R markdown file is 
provided with this package}

\item{output_file}{Output name (I recommend ending in 'html', as this is what the file
is) of your output file.}

\item{output_directory}{Directory the output file will be written to. 
Defaults to your working directory.}

\item{GEMINI_data}{Path to .Rdata data frame which contains the Data Frame 
of the GEMINI output that will be plotted. Helper scripts are provided as 
\code{\link{gemini_test_wrapper}} and \code{\link{gemini_query_wrapper}} 
which will return the GEMINI output as a data frame into your R session.}

\item{sample_name}{The name of your sample}

\item{title}{The title of the document}

\item{peddy_path_prefix}{Path and prefix for the peddy output}

\item{peddy_id}{A character vector of the samples you want to highlight in the
peddy QC tab}

\item{decorate}{TRUE or FALSE. If set to TRUE (default), then the data frame 
given as input `GEMINI_data` will be decorated with the default settings for
\code{\link{See_GEM_formatter}}. If you have already run this function on
the data frame you are giving to \code{\link{knit_see_gem}}, then set this
to FALSE.}

\item{skip_stats}{If set to 'yes' this will use an alternate template which
has no `peddy QC` tab.}
}
\value{
None
}
\description{
Create the interactive html document
}
\examples{
\dontrun{
# will output just the example document to ~/SeeGEM_document.html
knit_see_gem()
# more realistic example which does an automsal recessive test 
knit_see_gem(GEMINI_data = 
gemini_test_wrapper('2018_06_28__OGVFB_exomes.GATK.PED_master.gemini.db', 
test = 'autosomal_recessive', 
families = 'DDL003'), 
output_file='~/quick_SeeGEM.html', 
skip_stats = 'yes')
}
}
