% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_see_gem_from_table.R
\name{knit_see_gem_from_table}
\alias{knit_see_gem_from_table}
\title{Knit SeeGEM from tidy data}
\usage{

  knit_see_gem_from_table(rmd = system.file("rmd/document_template_from_table.Rmd",
  package = "SeeGEM"), output_file = "SeeGEM_document.html",
  output_directory = getwd(), table_data, sample_name = NA,
  title = "SeeGEM Test Report")
}
\arguments{
\item{rmd}{Path to a custom R markdown file. A default R markdown file is 
provided with this package}

\item{output_file}{Output name (I recommend ending in 'html', as this is what the file
is) of your output file.}

\item{output_directory}{Directory the output file will be written to. 
Defaults to your working directory.}

\item{table_data}{Data frame (or tibble) to be knit}

\item{sample_name}{The name of your sample}

\item{title}{The title of the document}
}
\value{
None
}
\description{
Create the interactive html document from a data frame / tibble
}
\examples{
\dontrun{
# will output just the example document to ~/SeeGEM_document.html
knit_see_gem()
# more realistic example which does an automsal recessive test 
knit_see_gem(GEMINI_data = 
gemini_test_wrapper('2018_06_28__OGVFB_exomes.GATK.PED_master.gemini.db', 
test = 'autosomal_recessive', 
families = 'DDL003'), 
output_file='~/quick_SeeGEM.html', 
skip_stats = 'yes')
}
}
