% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_generator.R
\name{link_generator}
\alias{link_generator}
\title{Link generator for See GEM interactive report}
\usage{
link_generator(base_url, ID, split_on = ",", link_name = NA)
}
\arguments{
\item{base_url}{The consistent base URL}

\item{ID}{The column ID to build the unique part of the URL}

\item{split_on}{What delimiter to split the field on? Defaults to comma.}

\item{link_name}{What to use for the link name? Defaults to using the row value of the column}
}
\value{
None
}
\description{
Takes a base url and a column from a data frame to generate a vector of URLs. 
You must use html format to specify special characters and spaces. 
Used in See_GEM_formatter function to format specified colums.
}
\examples{
link_generator('https://www.ncbi.nlm.nih.gov/projects/SNP/snp_ref.cgi?rs=', 'rs1800728')
}
