% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/See_GEM_formatter.R
\name{See_GEM_formatter}
\alias{See_GEM_formatter}
\title{GEMINI data formatter}
\usage{
See_GEM_formatter(GEMINI_data, core_fields = c("test", "pos_id", "gene",
  "impact_so", "hgvsc", "hgvsp", "aaf", "gno_af_all", "exac_num_hom_alt",
  "clinvar_id", "rs_ids", "GoogleScholar"), linkify = "yes",
  underscore_to_space = "yes")
}
\arguments{
\item{GEMINI_data}{The data frame / tibble of your GEMINI data. All tests 
(e.g. autosomal_recessive, compound_het) collapsed into a single data frame 
with the GEMINI sub command test given in a `test` column.}

\item{core_fields}{These are the columns that will be shown by default}

\item{linkify}{Do you want to turn format fields as hyperlinks? Hard-coded to 
position id (gnomAD), gene (OMIM), ClinVar ID (ClinVar), rs_id (dbSNP).}

\item{underscore_to_space}{Do you want to replace underscores with spaces?}
}
\value{
None
}
\description{
Formats data for reactive knit document
}
\examples{
GEMINI_data <- data.table::rbindlist(gemini_caller('/path/to/gemini.db', 'autosomal_recessive'),
gemini_caller('/path/to/gemini.db', 'autosomal_dominant', "aaf < 0.05 AND aaf_esp_all < 0.001 AND aaf_1kg_all < 0.001 AND af_exac_all < 0.001 AND (is_coding=1 OR is_splicing=1 OR impact_severity='HIGH') AND filter is NULL"))
See_GEM_formatter(GEMINI_data)
}
