% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_tibble.R
\name{vcf_to_tibble}
\alias{vcf_to_tibble}
\title{VCF to tibble}
\usage{
vcf_to_tibble(VCF)
}
\arguments{
\item{VCF}{Path to a VCF}
}
\value{
None
}
\description{
Creates a tidy tibble from a VCF file
}
\examples{

vcf_to_tibble(system.file("extdata/example.vcf.gz", package="SeeGEM"))

}
