% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/See_GEM_formatter.R
\name{See_GEM_formatter}
\alias{See_GEM_formatter}
\title{GEMINI data formatter}
\usage{
See_GEM_formatter(GEMINI_data, core_fields = c("test", "pos_id",
  "impact_so", "gene", "hgvsc", "hgvsp", "aaf", "gno_af_all",
  "exac_num_hom_alt", "clinvar_id", "rs_ids", "GoogleScholar"),
  in_silico = c("test", "pos_id", "gene", "impact_so", "cadd_phred",
  "ccr_pct_v1", "revel", "polyphen_score", "sift_score",
  "metalr_rankscore", "genesplicer", "spliceregion", "linsight"),
  genotypes = c("test", "pos_id", "gene", "impact_so", "family_members",
  "family_genotypes"),
  extra_columns_to_retain = "^gno|rankscore$|*num*|^clin|*domain*|*codon*",
  linkify = "yes", underscore_to_space = "yes")
}
\arguments{
\item{GEMINI_data}{The data frame / tibble of your GEMINI data. All tests 
(e.g. autosomal_recessive, compound_het) collapsed into a single data frame 
with the GEMINI sub command test given in a `test` column.}

\item{core_fields}{These are the columns that will be shown by default}

\item{in_silico}{These are in silico consequence columns (e.g CADD, SIFT)}

\item{genotypes}{These columns show the full genotype information for each person
in the family.}

\item{extra_columns_to_retain}{These are columns not selected above that will be 
retained. This is a regular expression, so you can give this something like 
'^gno|rankscore$|*num*' and all columns starting with `gno` or ending with `rankscore`
or containing `num` will be selected.}

\item{linkify}{Do you want to turn format fields as hyperlinks? Hard-coded to 
position id (gnomAD), gene (OMIM), ClinVar ID (ClinVar), rs_id (dbSNP).}

\item{underscore_to_space}{Do you want to replace underscores with spaces?}
}
\value{
None
}
\description{
Formats data for reactive knit document made with \code{\link{knit_see_gem}}
}
\examples{
\dontrun{
GEMINI_data <- data.table::rbindlist(gemini_test_wrapper('/path/to/gemini.db', 
'autosomal_recessive', families = 'the_fam'),
gemini_test_wrapper('/path/to/gemini.db', 'autosomal_dominant', filter = 
"aaf < 0.05 AND aaf_esp_all < 0.001 AND aaf_1kg_all < 0.001 AND 
af_exac_all < 0.001 AND 
(is_coding=1 OR is_splicing=1 OR impact_severity='HIGH') AND 
filter is NULL" , families = 'the_fam'))
See_GEM_formatter(GEMINI_data)
}
}
